6. Q&A Application on private documents/requirement.txt
openai
langchain
pinecone-client
python-dotenv
tiktoken
llmx
pypdf 
docx2txt 
streamlit
chromadb